solved bug: http://hub.qgis.org/issues/8934 due to porting error

+++ mirrorMap.py
@@ -62,7 +62,7 @@ class MirrorMap(QWidget):
 		self.canvas.enableAntiAliasing( settings.value( "/qgis/enable_anti_aliasing", False, type=bool ))
 		self.canvas.useImageToRender( settings.value( "/qgis/use_qimage_to_render", False, type=bool ))
 		action = settings.value( "/qgis/wheel_action", 0, type=int)
-		zoomFactor = settings.value( "/qgis/zoom_factor", 2, type=int )
+		zoomFactor = settings.value( "/qgis/zoom_factor", 2.0, type=float )
 		self.canvas.setWheelAction( QgsMapCanvas.WheelAction(action), zoomFactor )
 		gridLayout.addWidget( self.canvas, 0, 0, 1, 3 )
